pyproject.toml
[project]
name = "flowbothd"
version = "0.1.0"
description = "FlowBotHD: History-Aware Diffusion Handling Ambiguities in Articulated Objects Manipulation"
readme = "README.md"
requires-python = ">=3.9.7,<3.10"
license = { file = "LICENSE.txt" }
authors = [{ email = "yishul@andrew.cmu.edu", name = "Yishu Li" }]
dependencies = [
    "diffusers==0.19.3",
    "flowbot3d>=0.1.0",
    "huggingface==0.0.1",
    "huggingface-hub==0.16.4",
    "hydra-core==1.3.2",
    "lightning==2.0.3",
    "pyg-lib>=0.4.0",
    "rpad-core",
    "rpad-pybullet-envs>=0.1.0",
    "rpad-pybullet-libs>=0.1.0",
    "rpad-pyg>=0.1.0",
    "rpad-visualize-3d>=0.1.0",
    "timm==0.9.16",
    "torch==2.0.1",
    "torch-cluster==1.6.1+pt20cu118",
    "torch-geometric==2.2.0",
    "torch-scatter==2.1.1+pt20cu118",
    "torch-sparse==0.6.17+pt20cu118",
    "torch-spline-conv==1.2.2+pt20cu118",
    "torchaudio==2.0.2",
    "torchvision==0.15.2",
    "wandb==0.15.4",
]

[build-system]
requires = ["setuptools >= 58.0.1", "setuptools-scm", "wheel"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
develop = [
  "autoflake == 2.1.1",
  "black == 23.3.0",
  "isort == 5.12.0",
  "mypy == 1.3.0",
  "pandas-stubs == 2.0.1.230501",
  "pylint == 2.17.4",
  "pytest == 7.3.2",
  "pre-commit == 3.3.3",
]
notebooks = ["jupyter"]
build_docs = ["mkdocs-material", "mkdocstrings[python]"]

[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true


[tool.uv]
# these packages use torch, so we need to build them in the same environment
no-build-isolation-package = ["pyg-lib", "torch-scatter", "torch-sparse", "torch-cluster", "torch-spline-conv"]
find-links = [
  "https://data.pyg.org/whl/torch-2.0.1+cu118.html",
  "https://data.pyg.org/whl/torch-2.0.1+cpu.html",
]
# Override the dependency for partnet_mobility_utils because of a specific conflict
# in the branch we need compared to flowbot3d repo.
override-dependencies = ["rpad-partnet-mobility-utils"]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu118", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu118", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
rpad-core = { git = "https://github.com/r-pad/core.git" }
rpad-pybullet-envs = { git = "https://github.com/r-pad/pybullet_envs.git" }
rpad-pybullet-libs = { git = "https://github.com/r-pad/pybullet_libs.git" }
rpad-pyg = { git = "https://github.com/r-pad/pyg_libs.git" }
rpad-visualize-3d = { git = "https://github.com/r-pad/visualize_3d.git" }
flowbot3d = { git = "https://github.com/r-pad/flowbot3d.git" }
rpad-partnet-mobility-utils = { git = "https://github.com/r-pad/partnet_mobility_utils.git", branch = "diffusion" }


# This is required to allow us to have notebooks/ at the top level.
[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
flowbothd = ["py.typed"]

[tool.isort]
profile = "black"
known_third_party = "wandb"

[tool.mypy]
python_version = 3.9
warn_return_any = true
warn_unused_configs = true
mypy_path = "src"
namespace_packages = true
explicit_package_bases = true

[[tool.mypy.overrides]]
module = ["torchvision.*"]
ignore_missing_imports = true

[tool.pylint]
known-third-party = "wandb"

[tool.pylint.TYPECHECK]
generated-members = 'torch.*'
